+++ src/collective/fingerpointing/subscribers/pas_logger.py
@@ -9,7 +9,13 @@ from Products.PluggableAuthService.interfaces.events import IPrincipalCreatedEve
 from Products.PluggableAuthService.interfaces.events import IPrincipalDeletedEvent  # noqa: E501
 from Products.PluggableAuthService.interfaces.events import IUserLoggedInEvent
 from Products.PluggableAuthService.interfaces.events import IUserLoggedOutEvent
-from zope.component.interfaces import ComponentLookupError
+
+try:
+    # Plone 5.1+
+    from zope.interface.interfaces import ComponentLookupError
+except ImportError
+    # BBB Plone 5.0-
+    from zope.component.interfaces import ComponentLookupError
 
 
 def pas_logger(event):